requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.9
django-admin-berry==1.0.10
django-phonenumber-field==8.1.0
gunicorn==21.2.0
idna==3.10
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
python-dotenv==1.0.1
razorpay==1.4.2
requests==2.32.4
sqlparse==0.5.3
str2bool==1.1
typing_extensions==4.14.0
urllib3==2.4.0
whitenoise==6.6.0
